requirements.txt
bleach==6.0.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
cryptography==41.0.1
docutils==0.20.1
dotty-dict==1.3.1
Flask==2.3.3
Flask-JWT-Extended==4.5.2
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
invoke==2.1.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
keyring==24.0.0
MarkupSafe==2.1.3
more-itertools==9.1.0
packaging==23.1
pkginfo==1.9.6
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-gitlab==3.15.0
python-semantic-release==7.34.6
readme-renderer==40.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
SecretStorage==3.3.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.15
tomlkit==0.11.8
tqdm==4.65.0
twine==3.8.0
typing_extensions==4.6.3
urllib3==2.0.3
webencodings==0.5.1
Werkzeug==2.3.7
zipp==3.15.0
